# api/swagger/swagger.yaml
# Copyright 2024 costalong <costa9293@gmail.com>. All rights reserved.
# Use of this source code is governed by a MIT style
# license that can be found in the LICENSE file. The original repo for
# this file is https://github.com/costa92/micros-service

consumes:
    - application/json
produces:
    - application/json
swagger: "2.0"
info:
    title: order-server/v1/errors.proto
    version: version not set
paths:
    /v1/order:
        post:
            operationId: OrderServer_CreateOrder
            parameters:
                - in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/v1CreateOrderRequest'
            responses:
                "200":
                    description: A successful response.
                    schema:
                        $ref: '#/definitions/v1CreateOrderResponse'
                default:
                    description: An unexpected error response.
                    schema:
                        $ref: '#/definitions/rpcStatus'
            tags:
                - OrderServer
    /v1/order/{order_id}:
        get:
            operationId: OrderServer_Detail
            parameters:
                - in: path
                  name: order_id
                  required: true
                  type: string
            responses:
                "200":
                    description: A successful response.
                    schema:
                        $ref: '#/definitions/order_serverv1DetailResponse'
                default:
                    description: An unexpected error response.
                    schema:
                        $ref: '#/definitions/rpcStatus'
            tags:
                - OrderServer
    /v1/pay:
        post:
            operationId: PayServer_Pay
            parameters:
                - in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/v1PayRequest'
            responses:
                "200":
                    description: A successful response.
                    schema:
                        $ref: '#/definitions/v1PayResponse'
                default:
                    description: An unexpected error response.
                    schema:
                        $ref: '#/definitions/rpcStatus'
            tags:
                - PayServer
    /v1/pay/{order_id}:
        get:
            operationId: PayServer_Detail
            parameters:
                - in: path
                  name: order_id
                  required: true
                  type: string
            responses:
                "200":
                    description: A successful response.
                    schema:
                        $ref: '#/definitions/pay_serverv1DetailResponse'
                default:
                    description: An unexpected error response.
                    schema:
                        $ref: '#/definitions/rpcStatus'
            tags:
                - PayServer
definitions:
    order_serverv1DetailResponse:
        properties:
            amount:
                type: string
                x-order: !!float 2
            order_id:
                type: string
                x-order: !!float 0
            status:
                type: string
                x-order: !!float 3
            user_id:
                type: string
                x-order: !!float 1
        type: object
    pay_serverv1DetailResponse:
        properties:
            amount:
                type: string
                x-order: !!float 2
            order_id:
                type: string
                x-order: !!float 0
            status:
                type: string
                x-order: !!float 3
            user_id:
                type: string
                x-order: !!float 1
        type: object
    protobufAny:
        additionalProperties: {}
        properties:
            '@type':
                type: string
        type: object
    rpcStatus:
        properties:
            code:
                format: int32
                type: integer
            details:
                items:
                    $ref: '#/definitions/protobufAny'
                    type: object
                type: array
            message:
                type: string
        type: object
    v1CreateOrderRequest:
        properties:
            amount:
                type: string
                x-order: !!float 1
            user_id:
                type: string
                x-order: !!float 0
        type: object
    v1CreateOrderResponse:
        properties:
            amount:
                type: string
                x-order: !!float 2
            order_id:
                type: string
                x-order: !!float 0
            status:
                type: string
                x-order: !!float 3
            user_id:
                type: string
                x-order: !!float 1
        type: object
    v1PayRequest:
        properties:
            amount:
                type: string
                x-order: !!float 2
            order_id:
                type: string
                x-order: !!float 0
            user_id:
                type: string
                x-order: !!float 1
        type: object
    v1PayResponse:
        properties:
            amount:
                type: string
                x-order: !!float 2
            order_id:
                type: string
                x-order: !!float 0
            status:
                type: string
                x-order: !!float 3
            user_id:
                type: string
                x-order: !!float 1
        type: object
tags:
    - name: OrderServer
    - name: PayServer