# pkg/api/payserver/v1/pay-serve_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.5.3
// - protoc             v5.28.3
// source: payserver/v1/pay-serve.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationPayServerDetail = "/pay_server.v1.PayServer/Detail"
const OperationPayServerPay = "/pay_server.v1.PayServer/Pay"

type PayServerHTTPServer interface {
	Detail(context.Context, *DetailRequest) (*DetailResponse, error)
	Pay(context.Context, *PayRequest) (*PayResponse, error)
}

func RegisterPayServerHTTPServer(s *http.Server, srv PayServerHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/pay", _PayServer_Pay0_HTTP_Handler(srv))
	r.GET("/v1/pay/{order_id}", _PayServer_Detail1_HTTP_Handler(srv))
}

func _PayServer_Pay0_HTTP_Handler(srv PayServerHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PayRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPayServerPay)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Pay(ctx, req.(*PayRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PayResponse)
		return ctx.Result(200, reply)
	}
}

func _PayServer_Detail1_HTTP_Handler(srv PayServerHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DetailRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationPayServerDetail)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Detail(ctx, req.(*DetailRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DetailResponse)
		return ctx.Result(200, reply)
	}
}

type PayServerHTTPClient interface {
	Detail(ctx context.Context, req *DetailRequest, opts ...http.CallOption) (rsp *DetailResponse, err error)
	Pay(ctx context.Context, req *PayRequest, opts ...http.CallOption) (rsp *PayResponse, err error)
}

type PayServerHTTPClientImpl struct {
	cc *http.Client
}

func NewPayServerHTTPClient(client *http.Client) PayServerHTTPClient {
	return &PayServerHTTPClientImpl{client}
}

func (c *PayServerHTTPClientImpl) Detail(ctx context.Context, in *DetailRequest, opts ...http.CallOption) (*DetailResponse, error) {
	var out DetailResponse
	pattern := "/v1/pay/{order_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationPayServerDetail))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *PayServerHTTPClientImpl) Pay(ctx context.Context, in *PayRequest, opts ...http.CallOption) (*PayResponse, error) {
	var out PayResponse
	pattern := "/v1/pay"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationPayServerPay))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}